chore: Rename to OrderTradable trait and remove profitability calculator
